Cleanup config
