+++ backend/shop-templates/README.md
@@ -1,0 +1,3 @@
+Directory containing shop configuration templates and default configuration files.
+
+Used by the back-end logic during a new shop creation.
