+++ backend/shop-templates/affiliate/config.json
