+++ backend/shop-templates/baseConfig.js
