+++ backend/shop-templates/empty/config.json
