+++ backend/shop-templates/multi-product/collections.json
