+++ backend/shop-templates/multi-product/config.json
