+++ backend/shop-templates/multi-product/products.json
