+++ backend/shop-templates/multi-product/style.css
