+++ backend/shop-templates/single-product/config.json
