+++ backend/shop-templates/single-product/style.css
