+++ backend/utils/shop.js
@@ -3,75 +3,6 @@ const fs = require('fs')
 const { Shop } = require('../models')
 const { DSHOP_CACHE } = require('./const')
 
-/**
- * Create a new shop in the DB.
- *
- * @param {number} networkId: 1=Mainnet, 4=Rinkeby, 999=localhost, etc...
- * @param {string} name: shop name
- * @param {string} listingId: listing ID associated with the shop
- * @param {string} authToken: token
- * @param {string} config: Shop's encrypted configuration
- * @param {number} sellerId: If of the row in the SellerShop table for the shop's owner
- * @param {string} hostname: hostname of the shop URL
- * @returns {Promise<{shop: models.Shop}|{error: string, status: number}>}
- */
-async function createShop({
-  networkId,
-  name,
-  listingId,
-  authToken,
-  config,
-  sellerId,
-  hostname
-}) {
-  if (!name) {
-    return { status: 400, error: 'Provide a shop name' }
-  }
-  // Remove any leading/trailing space.
-  name = name.trim()
-
-  // Store name must only contain alpha-numeric characters and space.
-  // TODO: Add support for UTF_8 characters. This requires changes throughout the stack
-  //       and in particular encoding/decoding shop name in URLs.
-  if (!name.match(/^[a-zA-Z0-9\-' ]+$/)) {
-    return {
-      status: 400,
-      error:
-        'The shop name contains invalid character. Only alphabetical characters, numbers, space, hyphen and apostrophe are allowed.'
-    }
-  }
-  if (listingId && !String(listingId).match(/^[0-9]+-[0-9]+-[0-9]+$/)) {
-    return {
-      status: 400,
-      error: 'Listing ID must be of form xxx-xxx-xxx eg 1-001-123'
-    }
-  }
-  if (listingId && !listingId.startsWith(networkId)) {
-    return {
-      status: 400,
-      error: `Listing ID ${listingId} is not on expected Network ID ${networkId}`
-    }
-  }
-  if (!authToken) {
-    return { status: 400, error: 'Provide an auth token' }
-  }
-  if (!sellerId) {
-    return { status: 400, error: 'Provide a seller ID' }
-  }
-
-  const shop = await Shop.create({
-    name,
-    networkId,
-    listingId,
-    authToken,
-    config,
-    sellerId,
-    hostname
-  })
-
-  return { shop }
-}
-
 function findShopByHostname(req, res, next) {
   Shop.findOne({ where: { hostname: req.hostname } }).then((shop) => {
     req.shop = shop
@@ -146,7 +77,6 @@ function getShopDataUrl(shop, networkConfig) {
   )
 }
 
-
 async function _tryDataDir(dataDir) {
   const hasDir = fs.existsSync(`${DSHOP_CACHE}/${dataDir}`)
   const [authToken, hostname] = [dataDir, dataDir]
@@ -185,7 +115,6 @@ async function getDataDir(seedDataDir) {
 }
 
 module.exports = {
-  createShop,
   findShop,
   findShopByHostname,
   getShopDataUrl,
